Extract sharder classes of Sharder.scala
