+++ processor/src/main/scala/be/cetic/tsorage/processor/sharder/TimeFormatSharder.scala
@@ -1,0 +1,9 @@
+package be.cetic.tsorage.processor.sharder
+
+import java.time.LocalDateTime
+import java.time.format.DateTimeFormatter
+
+abstract class TimeFormatSharder(val formatter: DateTimeFormatter) extends Sharder
+{
+  def shard(dt: LocalDateTime) = dt.format(formatter)
+}
